Extract folder workflow and remove legacy group code
