+++ Localization/strings.ar.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "تم إنشاء اللقطة لـ \"{0}\".",
   "Projects.Page.SearchWatermark": "مشاريع البحث...",
   "Projects.Page.Title": "جميع المشاريع",
-  "Projects.Group.All": "جميع المشاريع",
-  "Projects.Group.Archive": "الأرشيف",
-  "Projects.Group.Critical": "حرج",
-  "Projects.Group.Games": "ألعاب",
-  "Projects.Group.Label": "العرض",
-  "Projects.Group.Help": "اختر عرضًا ثم نفذ الإجراءات لجميع المشاريع في تلك المجموعة.",
-  "Projects.Group.TagLabel": "وسم المجموعة",
-  "Projects.Group.TagWatermark": "اكتب أو اختر وسمًا موجودًا...",
-  "Projects.Group.TagApply": "تطبيق وسم",
-  "Projects.Group.TagRemove": "إزالة الوسم",
-  "Projects.Group.TagHelp": "ينطبق على جميع المشاريع المسجلة في العرض الحالي.",
-  "Projects.Group.TagApplied": "تم تطبيق الوسم '{0}' على {1} مشروع.",
-  "Projects.Group.TagRemoved": "تمت إزالة الوسم '{0}' من {1} مشروع.",
-  "Projects.Group.Media": "وسائط",
+  "Projects.Tags.Remove": "إزالة الوسم",
   "Projects.Group.BackupAction": "نسخ احتياطي للعرض المحدد",
   "Projects.Group.BackupQueued": "تمت جدولة النسخ الاحتياطي لـ {0} مشروع.",
   "Projects.Group.DisableAuto": "إيقاف النسخ الاحتياطي التلقائي",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "فشل تشغيل عملية اللقطات المجمعة: {0}",
   "Projects.Group.SnapshotFailure": "فشل إنشاء لقطات لـ {0} مشروع.",
   "Projects.Group.SnapshotSuccess": "تم إنشاء لقطات لـ {0} مشروع.",
-  "Projects.Group.Work": "عمل",
   "Projects.Preset.DescriptionFallback": "يتم تطبيق قواعد الإعداد المسبق من أنماط .vaultsyncignore.",
   "Projects.Preset.DescriptionUnknown": "الإعداد المسبق '{0}' نشط.",
   "Projects.Preset.ExampleLabel": "مثال:",
