+++ Localization/strings.bn.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "'{0}'-এর জন্য স্ন্যাপশট তৈরি হয়েছে।",
   "Projects.Page.SearchWatermark": "প্রকল্প অনুসন্ধান...",
   "Projects.Page.Title": "সব প্রকল্প",
-  "Projects.Group.All": "সব প্রকল্প",
-  "Projects.Group.Archive": "আর্কাইভ",
-  "Projects.Group.Critical": "গুরুত্বপূর্ণ",
-  "Projects.Group.Games": "গেমস",
-  "Projects.Group.Label": "ভিউ",
-  "Projects.Group.Help": "একটি ভিউ নির্বাচন করুন, তারপর সেই গ্রুপের সব প্রকল্পে অ্যাকশন চালান।",
-  "Projects.Group.TagLabel": "গ্রুপ ট্যাগ",
-  "Projects.Group.TagWatermark": "একটি বিদ্যমান ট্যাগ লিখুন বা নির্বাচন করুন...",
-  "Projects.Group.TagApply": "ট্যাগ প্রয়োগ করুন",
-  "Projects.Group.TagRemove": "ট্যাগ সরান",
-  "Projects.Group.TagHelp": "বর্তমান ভিউয়ের সব নিবন্ধিত প্রকল্পে প্রযোজ্য।",
-  "Projects.Group.TagApplied": "'{0}' ট্যাগ {1} প্রকল্পে প্রয়োগ করা হয়েছে।",
-  "Projects.Group.TagRemoved": "'{0}' ট্যাগ {1} প্রকল্প থেকে সরানো হয়েছে।",
-  "Projects.Group.Media": "মিডিয়া",
+  "Projects.Tags.Remove": "ট্যাগ সরান",
   "Projects.Group.BackupAction": "নির্বাচিত ভিউ ব্যাকআপ করুন",
   "Projects.Group.BackupQueued": "{0} প্রকল্পের জন্য ব্যাকআপ কিউ করা হয়েছে।",
   "Projects.Group.DisableAuto": "স্বয়ংক্রিয় ব্যাকআপ বন্ধ করুন",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "গ্রুপ স্ন্যাপশট অপারেশন চালাতে ব্যর্থ: {0}",
   "Projects.Group.SnapshotFailure": "{0} প্রকল্পের জন্য স্ন্যাপশট তৈরি ব্যর্থ হয়েছে।",
   "Projects.Group.SnapshotSuccess": "{0} প্রকল্পের জন্য স্ন্যাপশট তৈরি হয়েছে।",
-  "Projects.Group.Work": "কাজ",
   "Projects.Preset.DescriptionFallback": ".vaultsyncignore প্যাটার্ন থেকে প্রিসেট নিয়ম প্রয়োগ করা হয়।",
   "Projects.Preset.DescriptionUnknown": "প্রিসেট '{0}' সক্রিয় রয়েছে।",
   "Projects.Preset.ExampleLabel": "উদাহরণ:",
