+++ Localization/strings.de.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "Snapshot für '{0}' erstellt.",
   "Projects.Page.SearchWatermark": "Projekte suchen...",
   "Projects.Page.Title": "Alle Projekte",
-  "Projects.Group.All": "Alle Projekte",
-  "Projects.Group.Archive": "Archiv",
-  "Projects.Group.Critical": "Kritisch",
-  "Projects.Group.Games": "Spiele",
-  "Projects.Group.Label": "Ansicht",
-  "Projects.Group.Help": "Wählen Sie eine Ansicht und führen Sie dann Aktionen für alle Projekte in dieser Gruppe aus.",
-  "Projects.Group.TagLabel": "Gruppen-Tag",
-  "Projects.Group.TagWatermark": "Tag eingeben oder vorhandenen auswählen...",
-  "Projects.Group.TagApply": "Tag anwenden",
-  "Projects.Group.TagRemove": "Tag entfernen",
-  "Projects.Group.TagHelp": "Gilt für alle registrierten Projekte in der aktuellen Ansicht.",
-  "Projects.Group.TagApplied": "Tag '{0}' auf {1} Projekte angewendet.",
-  "Projects.Group.TagRemoved": "Tag '{0}' von {1} Projekten entfernt.",
-  "Projects.Group.Media": "Medien",
+  "Projects.Tags.Remove": "Tag entfernen",
   "Projects.Group.BackupAction": "Ausgewählte Ansicht sichern",
   "Projects.Group.BackupQueued": "Backup für {0} Projekte in Warteschlange gestellt.",
   "Projects.Group.DisableAuto": "Automatische Backups pausieren",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "Gruppierte Snapshot-Operation fehlgeschlagen: {0}",
   "Projects.Group.SnapshotFailure": "Snapshots für {0} Projekte konnten nicht erstellt werden.",
   "Projects.Group.SnapshotSuccess": "Snapshots für {0} Projekte erstellt.",
-  "Projects.Group.Work": "Arbeit",
   "Projects.Preset.DescriptionFallback": "Preset-Regeln werden aus .vaultsyncignore-Mustern angewendet.",
   "Projects.Preset.DescriptionUnknown": "Preset „{0}“ ist aktiv.",
   "Projects.Preset.ExampleLabel": "Beispiel:",
