+++ Localization/strings.en.json
@@ -795,20 +795,7 @@
   "Projects.Notification.SnapshotSuccess": "Snapshot created for '{0}'.",
   "Projects.Page.SearchWatermark": "Search projects...",
   "Projects.Page.Title": "All projects",
-  "Projects.Group.All": "All projects",
-  "Projects.Group.Archive": "Archive",
-  "Projects.Group.Critical": "Critical",
-  "Projects.Group.Games": "Games",
-  "Projects.Group.Label": "View",
-  "Projects.Group.Help": "Choose a view, then run actions for all projects in that group.",
-  "Projects.Group.TagLabel": "Group tag",
-  "Projects.Group.TagWatermark": "Type or pick an existing tag...",
-  "Projects.Group.TagApply": "Apply tag",
-  "Projects.Group.TagRemove": "Remove tag",
-  "Projects.Group.TagHelp": "Applies to all registered projects in the current view.",
-  "Projects.Group.TagApplied": "Applied tag '{0}' to {1} projects.",
-  "Projects.Group.TagRemoved": "Removed tag '{0}' from {1} projects.",
-  "Projects.Group.Media": "Media",
+  "Projects.Tags.Remove": "Remove tag",
   "Projects.Group.BackupAction": "Back up selected view",
   "Projects.Group.BackupQueued": "Queued backup for {0} projects.",
   "Projects.Group.DisableAuto": "Pause auto backups",
@@ -819,7 +806,6 @@
   "Projects.Group.SnapshotError": "Failed to run grouped snapshot operation: {0}",
   "Projects.Group.SnapshotFailure": "Failed to create snapshots for {0} projects.",
   "Projects.Group.SnapshotSuccess": "Created snapshots for {0} projects.",
-  "Projects.Group.Work": "Work",
   "Projects.Preset.DescriptionFallback": "Preset rules are applied from .vaultsyncignore patterns.",
   "Projects.Preset.DescriptionUnknown": "Preset '{0}' is active.",
   "Projects.Preset.ExampleLabel": "Example:",
