+++ Localization/strings.es.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "Instantánea creada para '{0}'.",
   "Projects.Page.SearchWatermark": "Buscar proyectos...",
   "Projects.Page.Title": "Todos los proyectos",
-  "Projects.Group.All": "Todos los proyectos",
-  "Projects.Group.Archive": "Archivo",
-  "Projects.Group.Critical": "Crítico",
-  "Projects.Group.Games": "Juegos",
-  "Projects.Group.Label": "Vista",
-  "Projects.Group.Help": "Elige una vista y ejecuta acciones para todos los proyectos de ese grupo.",
-  "Projects.Group.TagLabel": "Etiqueta de grupo",
-  "Projects.Group.TagWatermark": "Escribe o elige una etiqueta existente...",
-  "Projects.Group.TagApply": "Aplicar etiqueta",
-  "Projects.Group.TagRemove": "Eliminar etiqueta",
-  "Projects.Group.TagHelp": "Se aplica a todos los proyectos registrados en la vista actual.",
-  "Projects.Group.TagApplied": "Etiqueta '{0}' aplicada a {1} proyectos.",
-  "Projects.Group.TagRemoved": "Etiqueta '{0}' eliminada de {1} proyectos.",
-  "Projects.Group.Media": "Medios",
+  "Projects.Tags.Remove": "Eliminar etiqueta",
   "Projects.Group.BackupAction": "Respaldar vista seleccionada",
   "Projects.Group.BackupQueued": "Backup en cola para {0} proyectos.",
   "Projects.Group.DisableAuto": "Pausar backups automáticos",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "Error al ejecutar snapshot agrupado: {0}",
   "Projects.Group.SnapshotFailure": "No se pudieron crear snapshots para {0} proyectos.",
   "Projects.Group.SnapshotSuccess": "Snapshots creados para {0} proyectos.",
-  "Projects.Group.Work": "Trabajo",
   "Projects.Preset.DescriptionFallback": "Las reglas del preset se aplican desde patrones .vaultsyncignore.",
   "Projects.Preset.DescriptionUnknown": "El preset '{0}' está activo.",
   "Projects.Preset.ExampleLabel": "Ejemplo:",
