+++ Localization/strings.fr.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "Instantané créé pour '{0}'.",
   "Projects.Page.SearchWatermark": "Rechercher des projets...",
   "Projects.Page.Title": "Tous les projets",
-  "Projects.Group.All": "Tous les projets",
-  "Projects.Group.Archive": "Archive",
-  "Projects.Group.Critical": "Critique",
-  "Projects.Group.Games": "Jeux",
-  "Projects.Group.Label": "Vue",
-  "Projects.Group.Help": "Choisissez une vue puis exécutez des actions pour tous les projets de ce groupe.",
-  "Projects.Group.TagLabel": "Étiquette du groupe",
-  "Projects.Group.TagWatermark": "Tapez ou choisissez une étiquette existante...",
-  "Projects.Group.TagApply": "Appliquer une étiquette",
-  "Projects.Group.TagRemove": "Supprimer l'étiquette",
-  "Projects.Group.TagHelp": "S'applique à tous les projets enregistrés dans la vue actuelle.",
-  "Projects.Group.TagApplied": "Étiquette '{0}' appliquée à {1} projets.",
-  "Projects.Group.TagRemoved": "Étiquette '{0}' supprimée de {1} projets.",
-  "Projects.Group.Media": "Médias",
+  "Projects.Tags.Remove": "Supprimer l'étiquette",
   "Projects.Group.BackupAction": "Sauvegarder la vue sélectionnée",
   "Projects.Group.BackupQueued": "Sauvegarde mise en file d'attente pour {0} projets.",
   "Projects.Group.DisableAuto": "Mettre en pause les sauvegardes automatiques",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "Échec de l'opération de snapshot groupée : {0}",
   "Projects.Group.SnapshotFailure": "Impossible de créer des snapshots pour {0} projets.",
   "Projects.Group.SnapshotSuccess": "Snapshots créés pour {0} projets.",
-  "Projects.Group.Work": "Travail",
   "Projects.Preset.DescriptionFallback": "Les règles du preset sont appliquées depuis les modèles .vaultsyncignore.",
   "Projects.Preset.DescriptionUnknown": "Le preset '{0}' est actif.",
   "Projects.Preset.ExampleLabel": "Exemple:",
