+++ Localization/strings.hi.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "'{0}' के लिए स्नैपशॉट बनाया गया।",
   "Projects.Page.SearchWatermark": "प्रोजेक्ट खोजें...",
   "Projects.Page.Title": "सभी प्रोजेक्ट",
-  "Projects.Group.All": "सभी प्रोजेक्ट",
-  "Projects.Group.Archive": "आर्काइव",
-  "Projects.Group.Critical": "महत्वपूर्ण",
-  "Projects.Group.Games": "गेम्स",
-  "Projects.Group.Label": "दृश्य",
-  "Projects.Group.Help": "एक दृश्य चुनें, फिर उस समूह के सभी प्रोजेक्ट पर कार्य चलाएँ।",
-  "Projects.Group.TagLabel": "समूह टैग",
-  "Projects.Group.TagWatermark": "टैग लिखें या मौजूदा चुनें...",
-  "Projects.Group.TagApply": "टैग लागू करें",
-  "Projects.Group.TagRemove": "टैग हटाएँ",
-  "Projects.Group.TagHelp": "वर्तमान दृश्य के सभी पंजीकृत प्रोजेक्ट पर लागू होता है।",
-  "Projects.Group.TagApplied": "टैग '{0}' को {1} प्रोजेक्ट पर लागू किया गया।",
-  "Projects.Group.TagRemoved": "टैग '{0}' को {1} प्रोजेक्ट से हटाया गया।",
-  "Projects.Group.Media": "मीडिया",
+  "Projects.Tags.Remove": "टैग हटाएँ",
   "Projects.Group.BackupAction": "चयनित दृश्य का बैकअप लें",
   "Projects.Group.BackupQueued": "{0} प्रोजेक्ट के लिए बैकअप कतार में डाला गया।",
   "Projects.Group.DisableAuto": "स्वचालित बैकअप रोकें",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "समूह स्नैपशॉट ऑपरेशन विफल: {0}",
   "Projects.Group.SnapshotFailure": "{0} प्रोजेक्ट के लिए स्नैपशॉट बनाने में विफल।",
   "Projects.Group.SnapshotSuccess": "{0} प्रोजेक्ट के लिए स्नैपशॉट बनाए गए।",
-  "Projects.Group.Work": "काम",
   "Projects.Preset.DescriptionFallback": "प्रीसेट नियम .vaultsyncignore पैटर्न से लागू होते हैं।",
   "Projects.Preset.DescriptionUnknown": "प्रीसेट '{0}' सक्रिय है।",
   "Projects.Preset.ExampleLabel": "उदाहरण:",
