+++ Localization/strings.id.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "Snapshot dibuat untuk '{0}'.",
   "Projects.Page.SearchWatermark": "Cari proyek...",
   "Projects.Page.Title": "Semua proyek",
-  "Projects.Group.All": "Semua proyek",
-  "Projects.Group.Archive": "Arsip",
-  "Projects.Group.Critical": "Kritis",
-  "Projects.Group.Games": "Permainan",
-  "Projects.Group.Label": "Lihat",
-  "Projects.Group.Help": "Pilih tampilan, lalu jalankan tindakan untuk semua proyek dalam grup tersebut.",
-  "Projects.Group.TagLabel": "Tag grup",
-  "Projects.Group.TagWatermark": "Ketik atau pilih tag yang ada...",
-  "Projects.Group.TagApply": "Terapkan tag",
-  "Projects.Group.TagRemove": "Hapus tag",
-  "Projects.Group.TagHelp": "Berlaku untuk semua proyek terdaftar dalam tampilan saat ini.",
-  "Projects.Group.TagApplied": "Menerapkan tag '{0}' ke proyek {1}.",
-  "Projects.Group.TagRemoved": "Menghapus tag '{0}' dari proyek {1}.",
-  "Projects.Group.Media": "Media",
+  "Projects.Tags.Remove": "Hapus tag",
   "Projects.Group.BackupAction": "Mencadangkan tampilan yang dipilih",
   "Projects.Group.BackupQueued": "Pencadangan antrian untuk proyek {0}.",
   "Projects.Group.DisableAuto": "Jeda pencadangan otomatis",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "Gagal menjalankan operasi rekam jepret yang dikelompokkan: {0}",
   "Projects.Group.SnapshotFailure": "Gagal membuat rekam jepret untuk proyek {0}.",
   "Projects.Group.SnapshotSuccess": "Membuat rekam jepret untuk proyek {0}.",
-  "Projects.Group.Work": "Bekerja",
   "Projects.Preset.DescriptionFallback": "Aturan prasetel diterapkan dari pola .vaultsyncignore.",
   "Projects.Preset.DescriptionUnknown": "Preset '{0}' aktif.",
   "Projects.Preset.ExampleLabel": "Contoh:",
