+++ Localization/strings.it.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "Snapshot creato per '{0}'.",
   "Projects.Page.SearchWatermark": "Cerca progetti...",
   "Projects.Page.Title": "Tutti i progetti",
-  "Projects.Group.All": "Tutti i progetti",
-  "Projects.Group.Archive": "Archivio",
-  "Projects.Group.Critical": "Critico",
-  "Projects.Group.Games": "Giochi",
-  "Projects.Group.Label": "Vista",
-  "Projects.Group.Help": "Scegli una vista, poi esegui azioni per tutti i progetti di quel gruppo.",
-  "Projects.Group.TagLabel": "Tag gruppo",
-  "Projects.Group.TagWatermark": "Scrivi o scegli un tag esistente...",
-  "Projects.Group.TagApply": "Applica tag",
-  "Projects.Group.TagRemove": "Rimuovi tag",
-  "Projects.Group.TagHelp": "Si applica a tutti i progetti registrati nella vista corrente.",
-  "Projects.Group.TagApplied": "Tag '{0}' applicato a {1} progetti.",
-  "Projects.Group.TagRemoved": "Tag '{0}' rimosso da {1} progetti.",
-  "Projects.Group.Media": "Media",
+  "Projects.Tags.Remove": "Rimuovi tag",
   "Projects.Group.BackupAction": "Esegui backup della vista selezionata",
   "Projects.Group.BackupQueued": "Backup in coda per {0} progetti.",
   "Projects.Group.DisableAuto": "Metti in pausa backup automatici",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "Errore nell'operazione snapshot di gruppo: {0}",
   "Projects.Group.SnapshotFailure": "Impossibile creare snapshot per {0} progetti.",
   "Projects.Group.SnapshotSuccess": "Snapshot creati per {0} progetti.",
-  "Projects.Group.Work": "Lavoro",
   "Projects.Preset.DescriptionFallback": "Le regole del preset sono applicate dai pattern .vaultsyncignore.",
   "Projects.Preset.DescriptionUnknown": "Il preset '{0}' è attivo.",
   "Projects.Preset.ExampleLabel": "Esempio:",
