+++ Localization/strings.ja.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "「{0}」のためにスナップショットが作成されました。",
   "Projects.Page.SearchWatermark": "検索プロジェクト...",
   "Projects.Page.Title": "すべてのプロジェクト",
-  "Projects.Group.All": "すべてのプロジェクト",
-  "Projects.Group.Archive": "アーカイブ",
-  "Projects.Group.Critical": "重要な点",
-  "Projects.Group.Games": "ゲーム",
-  "Projects.Group.Label": "眺め",
-  "Projects.Group.Help": "ビューを選択し、そのグループ内のすべてのプロジェクトに対してアクションを実行します。",
-  "Projects.Group.TagLabel": "グループタグ",
-  "Projects.Group.TagWatermark": "既存のタグをタイプするか選ぶか...",
-  "Projects.Group.TagApply": "タグを適用",
-  "Projects.Group.TagRemove": "タグを削除",
-  "Projects.Group.TagHelp": "現在のビューで登録されているすべてのプロジェクトに適用されます。",
-  "Projects.Group.TagApplied": "{1}プロジェクトにタグ「{0}」を適用しました。",
-  "Projects.Group.TagRemoved": "{1}プロジェクトからタグ「{0}」を削除しました。",
-  "Projects.Group.Media": "メディア",
+  "Projects.Tags.Remove": "タグを削除",
   "Projects.Group.BackupAction": "選択したビューをバックアップ",
   "Projects.Group.BackupQueued": "{0}プロジェクトのバックアップを待機させました。",
   "Projects.Group.DisableAuto": "一時停止自動バックアップ",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "グループスナップショット操作の失敗:{0}",
   "Projects.Group.SnapshotFailure": "{0}プロジェクトのスナップショット作成に失敗しました。",
   "Projects.Group.SnapshotSuccess": "{0}プロジェクトのスナップショットを作成しました。",
-  "Projects.Group.Work": "業績",
   "Projects.Preset.DescriptionFallback": "プリセットルールは.vaultsyncignoreパターンから適用されます。",
   "Projects.Preset.DescriptionUnknown": "プリセットの「{0}」は有効です。",
   "Projects.Preset.ExampleLabel": "例:",
