+++ Localization/strings.ko.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "'{0}'를 위해 스냅샷이 생성되었습니다.",
   "Projects.Page.SearchWatermark": "검색 프로젝트...",
   "Projects.Page.Title": "모든 프로젝트",
-  "Projects.Group.All": "모든 프로젝트",
-  "Projects.Group.Archive": "아카이브",
-  "Projects.Group.Critical": "비평",
-  "Projects.Group.Games": "게임",
-  "Projects.Group.Label": "전망",
-  "Projects.Group.Help": "뷰를 선택한 후, 그 그룹 내 모든 프로젝트에 대해 액션을 실행하세요.",
-  "Projects.Group.TagLabel": "그룹 태그",
-  "Projects.Group.TagWatermark": "기존 태그를 타이핑하거나 선택하세요...",
-  "Projects.Group.TagApply": "태그 적용",
-  "Projects.Group.TagRemove": "태그 제거",
-  "Projects.Group.TagHelp": "현재 뷰에서 등록된 모든 프로젝트에 적용됩니다.",
-  "Projects.Group.TagApplied": "{1} 프로젝트에 '{0}' 태그를 적용했습니다.",
-  "Projects.Group.TagRemoved": "{1} 프로젝트에서 '{0}' 태그를 제거했습니다.",
-  "Projects.Group.Media": "미디어",
+  "Projects.Tags.Remove": "태그 제거",
   "Projects.Group.BackupAction": "선택 뷰 백업",
   "Projects.Group.BackupQueued": "{0} 프로젝트에 백업을 대기열에 넣었습니다.",
   "Projects.Group.DisableAuto": "자동 백업 일시정지",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "그룹 스냅샷 작업 실패: {0}",
   "Projects.Group.SnapshotFailure": "{0} 프로젝트에 대한 스냅샷을 생성하지 못했습니다.",
   "Projects.Group.SnapshotSuccess": "{0} 프로젝트의 스냅샷을 생성했습니다.",
-  "Projects.Group.Work": "작품",
   "Projects.Preset.DescriptionFallback": "프리셋 규칙은 .vaultsyncignore 패턴에서 적용됩니다.",
   "Projects.Preset.DescriptionUnknown": "프리셋 '{0}'이 활성화되어 있습니다.",
   "Projects.Preset.ExampleLabel": "예시:",
