+++ Localization/strings.nl.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "Snapshot gemaakt voor '{0}'.",
   "Projects.Page.SearchWatermark": "Zoekprojecten...",
   "Projects.Page.Title": "Alle projecten",
-  "Projects.Group.All": "Alle projecten",
-  "Projects.Group.Archive": "Archief",
-  "Projects.Group.Critical": "Kritisch",
-  "Projects.Group.Games": "Spellen",
-  "Projects.Group.Label": "Uitzicht",
-  "Projects.Group.Help": "Kies een weergave en voer acties uit voor alle projecten in die groep.",
-  "Projects.Group.TagLabel": "Groepstag",
-  "Projects.Group.TagWatermark": "Typ of kies een bestaand tag...",
-  "Projects.Group.TagApply": "Tag toepassen",
-  "Projects.Group.TagRemove": "Tag verwijderen",
-  "Projects.Group.TagHelp": "Van toepassing op alle geregistreerde projecten in de huidige weergave.",
-  "Projects.Group.TagApplied": "Label '{0}' toegepast op {1} projecten.",
-  "Projects.Group.TagRemoved": "Tag '{0}' uit {1} projecten verwijderd.",
-  "Projects.Group.Media": "Media",
+  "Projects.Tags.Remove": "Tag verwijderen",
   "Projects.Group.BackupAction": "Reserve-geselecteerde weergave",
   "Projects.Group.BackupQueued": "Backup voor {0} projecten in de wachtrij gezet.",
   "Projects.Group.DisableAuto": "Pauzeer automatische back-ups",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "Mislukt om een gegroepeerde snapshot-operatie uit te voeren: {0}",
   "Projects.Group.SnapshotFailure": "Het is niet gelukt om snapshots te maken voor {0} projecten.",
   "Projects.Group.SnapshotSuccess": "Snapshots gemaakt voor {0} projecten.",
-  "Projects.Group.Work": "Werk",
   "Projects.Preset.DescriptionFallback": "Preset-regels worden toegepast vanuit .vaultsyncignore-patronen.",
   "Projects.Preset.DescriptionUnknown": "Preset '{0}' is actief.",
   "Projects.Preset.ExampleLabel": "Voorbeeld:",
