+++ Localization/strings.pl.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "Snapshot stworzony dla '{0}'.",
   "Projects.Page.SearchWatermark": "Szukaj projektów...",
   "Projects.Page.Title": "Wszystkie projekty",
-  "Projects.Group.All": "Wszystkie projekty",
-  "Projects.Group.Archive": "Archiwum",
-  "Projects.Group.Critical": "Krytyczne",
-  "Projects.Group.Games": "Gry",
-  "Projects.Group.Label": "Zobacz",
-  "Projects.Group.Help": "Wybierz widok, a następnie wykonaj działania dla wszystkich projektów w tej grupie.",
-  "Projects.Group.TagLabel": "Tag grupowy",
-  "Projects.Group.TagWatermark": "Wpisz lub wybierz istniejący tag...",
-  "Projects.Group.TagApply": "Zastosuj tag",
-  "Projects.Group.TagRemove": "Usuń tag",
-  "Projects.Group.TagHelp": "Dotyczy wszystkich zarejestrowanych projektów w obecnym widoku.",
-  "Projects.Group.TagApplied": "Przypisałem tag '{0}' do {1} projektów.",
-  "Projects.Group.TagRemoved": "Usunięto tag '{0}' z {1} projektów.",
-  "Projects.Group.Media": "Media",
+  "Projects.Tags.Remove": "Usuń tag",
   "Projects.Group.BackupAction": "Backup wybranego widoku",
   "Projects.Group.BackupQueued": "Kolejka kopii zapasowych dla {0} projektów.",
   "Projects.Group.DisableAuto": "Wstrzymaj automatyczne kopie zapasowe",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "Nie udało się uruchomić operacji snapshotu grupowego: {0}",
   "Projects.Group.SnapshotFailure": "Nie udało się stworzyć migawek dla {0} projektów.",
   "Projects.Group.SnapshotSuccess": "Tworzyłem migawki do {0} projektów.",
-  "Projects.Group.Work": "Twórczość",
   "Projects.Preset.DescriptionFallback": "Reguły preset są stosowane na podstawie wzorców .vaultsyncignore.",
   "Projects.Preset.DescriptionUnknown": "Ustawienie '{0}' jest aktywne.",
   "Projects.Preset.ExampleLabel": "Przykład:",
