+++ Localization/strings.pt.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "Instantâneo criado para '{0}'.",
   "Projects.Page.SearchWatermark": "Buscar projetos...",
   "Projects.Page.Title": "Todos os projetos",
-  "Projects.Group.All": "Todos os projetos",
-  "Projects.Group.Archive": "Arquivo",
-  "Projects.Group.Critical": "Crítico",
-  "Projects.Group.Games": "Jogos",
-  "Projects.Group.Label": "Visualização",
-  "Projects.Group.Help": "Escolha uma visualização e execute ações para todos os projetos desse grupo.",
-  "Projects.Group.TagLabel": "Tag do grupo",
-  "Projects.Group.TagWatermark": "Digite ou escolha uma tag existente...",
-  "Projects.Group.TagApply": "Aplicar tag",
-  "Projects.Group.TagRemove": "Remover tag",
-  "Projects.Group.TagHelp": "Aplica-se a todos os projetos registrados na visualização atual.",
-  "Projects.Group.TagApplied": "Tag '{0}' aplicada a {1} projetos.",
-  "Projects.Group.TagRemoved": "Tag '{0}' removida de {1} projetos.",
-  "Projects.Group.Media": "Mídia",
+  "Projects.Tags.Remove": "Remover tag",
   "Projects.Group.BackupAction": "Fazer backup da visualização selecionada",
   "Projects.Group.BackupQueued": "Backup enfileirado para {0} projetos.",
   "Projects.Group.DisableAuto": "Pausar backups automáticos",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "Falha ao executar operação de snapshot em grupo: {0}",
   "Projects.Group.SnapshotFailure": "Falha ao criar snapshots para {0} projetos.",
   "Projects.Group.SnapshotSuccess": "Snapshots criados para {0} projetos.",
-  "Projects.Group.Work": "Trabalho",
   "Projects.Preset.DescriptionFallback": "As regras do preset são aplicadas a partir dos padrões .vaultsyncignore.",
   "Projects.Preset.DescriptionUnknown": "O preset '{0}' está ativo.",
   "Projects.Preset.ExampleLabel": "Exemplo:",
