+++ Localization/strings.ru.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "Снимок для проекта '{0}' успешно создан.",
   "Projects.Page.SearchWatermark": "Поиск проектов...",
   "Projects.Page.Title": "Все проекты",
-  "Projects.Group.All": "Все проекты",
-  "Projects.Group.Archive": "Архив",
-  "Projects.Group.Critical": "Критично",
-  "Projects.Group.Games": "Игры",
-  "Projects.Group.Label": "Вид",
-  "Projects.Group.Help": "Выберите вид, затем выполните действия для всех проектов этой группы.",
-  "Projects.Group.TagLabel": "Тег группы",
-  "Projects.Group.TagWatermark": "Введите или выберите существующий тег...",
-  "Projects.Group.TagApply": "Применить тег",
-  "Projects.Group.TagRemove": "Удалить тег",
-  "Projects.Group.TagHelp": "Применяется ко всем зарегистрированным проектам в текущем виде.",
-  "Projects.Group.TagApplied": "Тег '{0}' применён к {1} проектам.",
-  "Projects.Group.TagRemoved": "Тег '{0}' удалён из {1} проектов.",
-  "Projects.Group.Media": "Медиа",
+  "Projects.Tags.Remove": "Удалить тег",
   "Projects.Group.BackupAction": "Сделать резервную копию выбранного вида",
   "Projects.Group.BackupQueued": "Резервное копирование поставлено в очередь для {0} проектов.",
   "Projects.Group.DisableAuto": "Приостановить авто-бэкапы",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "Ошибка выполнения групповой операции снимка: {0}",
   "Projects.Group.SnapshotFailure": "Не удалось создать снимки для {0} проектов.",
   "Projects.Group.SnapshotSuccess": "Снимки созданы для {0} проектов.",
-  "Projects.Group.Work": "Работа",
   "Projects.Preset.DescriptionFallback": "Правила пресета применяются из шаблонов .vaultsyncignore.",
   "Projects.Preset.DescriptionUnknown": "Активен пресет '{0}'.",
   "Projects.Preset.ExampleLabel": "Пример:",
