+++ Localization/strings.tr.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "'{0}' için oluşturulan anlık görüntü.",
   "Projects.Page.SearchWatermark": "Projeleri arar...",
   "Projects.Page.Title": "Tüm projeler",
-  "Projects.Group.All": "Tüm projeler",
-  "Projects.Group.Archive": "Arşiv",
-  "Projects.Group.Critical": "Kritik",
-  "Projects.Group.Games": "Oyunlar",
-  "Projects.Group.Label": "Görünüm",
-  "Projects.Group.Help": "Bir görünüm seçin, sonra o gruptaki tüm projeler için eylemleri çalıştırın.",
-  "Projects.Group.TagLabel": "Grup etiketi",
-  "Projects.Group.TagWatermark": "Mevcut bir etiket yaz veya seç...",
-  "Projects.Group.TagApply": "Etiketi uygula",
-  "Projects.Group.TagRemove": "Etiketi kaldır",
-  "Projects.Group.TagHelp": "Mevcut görünümdeki tüm kayıtlı projeler için geçerlidir.",
-  "Projects.Group.TagApplied": "{1} projelerine '{0}' etiketi uyguladım.",
-  "Projects.Group.TagRemoved": "{1} projelerinden '{0}' etiketi kaldırıldı.",
-  "Projects.Group.Media": "Medya",
+  "Projects.Tags.Remove": "Etiketi kaldır",
   "Projects.Group.BackupAction": "Seçilmiş görünümü yedekleme",
   "Projects.Group.BackupQueued": "{0} projeleri için yedek sıraya girdim.",
   "Projects.Group.DisableAuto": "Otomatik yedeklemeleri durdur",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "Gruplanmış anlık görüntü işlemi çalıştırılamadı: {0}",
   "Projects.Group.SnapshotFailure": "{0} projeleri için anlık görüntüler oluşturamadı.",
   "Projects.Group.SnapshotSuccess": "{0} projeleri için anlık görüntüler oluşturdum.",
-  "Projects.Group.Work": "Çalışmaları",
   "Projects.Preset.DescriptionFallback": "Önceden belirlenmiş kurallar .vaultsyncignore desenlerinden uygulanır.",
   "Projects.Preset.DescriptionUnknown": "Önceden ayarlanmış '{0}' aktif.",
   "Projects.Preset.ExampleLabel": "Örnek:",
