+++ Localization/strings.uk.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "Знімок, створений для '{0}'.",
   "Projects.Page.SearchWatermark": "Пошукові проєкти...",
   "Projects.Page.Title": "Всі проєкти",
-  "Projects.Group.All": "Всі проєкти",
-  "Projects.Group.Archive": "Архів",
-  "Projects.Group.Critical": "Критично",
-  "Projects.Group.Games": "Ігри",
-  "Projects.Group.Label": "Погляд",
-  "Projects.Group.Help": "Виберіть вигляд, а потім запустіть дії для всіх проєктів у цій групі.",
-  "Projects.Group.TagLabel": "Тег групи",
-  "Projects.Group.TagWatermark": "Введіть або виберіть існуючий тег...",
-  "Projects.Group.TagApply": "Застосовувати тег",
-  "Projects.Group.TagRemove": "Видалити тег",
-  "Projects.Group.TagHelp": "Застосовується до всіх зареєстрованих проєктів у поточному вигляді.",
-  "Projects.Group.TagApplied": "Застосував тег '{0}' до {1} проєктів.",
-  "Projects.Group.TagRemoved": "Прибрано тег '{0}' з {1} проєктів.",
-  "Projects.Group.Media": "Медіа",
+  "Projects.Tags.Remove": "Видалити тег",
   "Projects.Group.BackupAction": "Резервне копіювання вибраного вигляду",
   "Projects.Group.BackupQueued": "Резервне копіювання в черзі для {0} проєктів.",
   "Projects.Group.DisableAuto": "Пауза автоматичного резервного копіювання",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "Не вдалося запустити згруповану операцію знімка: {0}",
   "Projects.Group.SnapshotFailure": "Не зміг створити знімки для {0} проєктів.",
   "Projects.Group.SnapshotSuccess": "Створював знімки для {0} проєктів.",
-  "Projects.Group.Work": "Творчість",
   "Projects.Preset.DescriptionFallback": "Правила заздалегідь встановлених застосовуються з шаблонів .vaultsyncignore (патернів .vaultsyncignore функцій).",
   "Projects.Preset.DescriptionUnknown": "Активний пресет '{0}'.",
   "Projects.Preset.ExampleLabel": "Приклад:",
