+++ Localization/strings.vi.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "Ảnh chụp nhanh được tạo cho '{0}'.",
   "Projects.Page.SearchWatermark": "Tìm kiếm dự án...",
   "Projects.Page.Title": "Tất cả các dự án",
-  "Projects.Group.All": "Tất cả các dự án",
-  "Projects.Group.Archive": "Lưu trữ",
-  "Projects.Group.Critical": "Quan trọng",
-  "Projects.Group.Games": "Trò chơi",
-  "Projects.Group.Label": "Xem",
-  "Projects.Group.Help": "Chọn một chế độ xem, sau đó chạy các tác vụ cho tất cả các dự án trong nhóm đó.",
-  "Projects.Group.TagLabel": "Thẻ nhóm",
-  "Projects.Group.TagWatermark": "Nhập hoặc chọn thẻ hiện có...",
-  "Projects.Group.TagApply": "Áp dụng thẻ",
-  "Projects.Group.TagRemove": "Xóa thẻ",
-  "Projects.Group.TagHelp": "Áp dụng cho tất cả các dự án đã đăng ký trong chế độ xem hiện tại.",
-  "Projects.Group.TagApplied": "Áp dụng thẻ '{0}' cho {1} dự án.",
-  "Projects.Group.TagRemoved": "Xóa thẻ '{0}' khỏi {1} dự án.",
-  "Projects.Group.Media": "Truyền thông",
+  "Projects.Tags.Remove": "Xóa thẻ",
   "Projects.Group.BackupAction": "Sao lưu chế độ xem đã chọn",
   "Projects.Group.BackupQueued": "Sao lưu hàng đợi cho {0} dự án.",
   "Projects.Group.DisableAuto": "Tạm dừng sao lưu tự động",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "Không chạy được thao tác chụp nhanh được nhóm: {0}",
   "Projects.Group.SnapshotFailure": "Không tạo được ảnh chụp nhanh cho {0} dự án.",
   "Projects.Group.SnapshotSuccess": "Tạo ảnh chụp nhanh cho {0} dự án.",
-  "Projects.Group.Work": "Công việc",
   "Projects.Preset.DescriptionFallback": "Các quy tắc đặt trước được áp dụng từ các mẫu .vaultsyncignore.",
   "Projects.Preset.DescriptionUnknown": "Cài đặt trước '{0}' đang hoạt động.",
   "Projects.Preset.ExampleLabel": "Ví dụ:",
