+++ Localization/strings.zh.json
@@ -819,20 +819,7 @@
   "Projects.Notification.SnapshotSuccess": "项目“{0}”的快照已成功创建。",
   "Projects.Page.SearchWatermark": "搜索项目…",
   "Projects.Page.Title": "所有项目",
-  "Projects.Group.All": "所有项目",
-  "Projects.Group.Archive": "归档",
-  "Projects.Group.Critical": "关键",
-  "Projects.Group.Games": "游戏",
-  "Projects.Group.Label": "视图",
-  "Projects.Group.Help": "选择一个视图，然后对该组中的所有项目执行操作。",
-  "Projects.Group.TagLabel": "分组标签",
-  "Projects.Group.TagWatermark": "输入或选择一个已有标签...",
-  "Projects.Group.TagApply": "应用标签",
-  "Projects.Group.TagRemove": "移除标签",
-  "Projects.Group.TagHelp": "适用于当前视图中的所有注册项目。",
-  "Projects.Group.TagApplied": "已将标签 '{0}' 应用于 {1} 个项目。",
-  "Projects.Group.TagRemoved": "已从 {1} 个项目中移除标签 '{0}'。",
-  "Projects.Group.Media": "媒体",
+  "Projects.Tags.Remove": "移除标签",
   "Projects.Group.BackupAction": "备份当前视图",
   "Projects.Group.BackupQueued": "已为 {0} 个项目加入备份队列。",
   "Projects.Group.DisableAuto": "暂停自动备份",
@@ -843,7 +830,6 @@
   "Projects.Group.SnapshotError": "执行分组快照操作失败：{0}",
   "Projects.Group.SnapshotFailure": "为 {0} 个项目创建快照失败。",
   "Projects.Group.SnapshotSuccess": "已为 {0} 个项目创建快照。",
-  "Projects.Group.Work": "工作",
   "Projects.Preset.DescriptionFallback": "预设规则来自 .vaultsyncignore 模式。",
   "Projects.Preset.DescriptionUnknown": "预设 '{0}' 已启用。",
   "Projects.Preset.ExampleLabel": "示例：",
